002/requirements.txt
azure-batch==14.0.0